+++ bizside_test_app/test/controllers/top_controller_test.rb
@@ -12,20 +12,20 @@ class TopControllerTest < ActionDispatch::IntegrationTest
     get '/', params: {ua: :pc}
     assert_response :success
     assert_template :"show"
-    assert response.body.include?('show.html.erb'), 'show.pc.html.erb は無いので show.html.erb を表示する'
+    assert response.body.include?('show.html.erb'), 'show.html+pc.erb は無いので show.html.erb を表示する'
   end
 
   def test_トップ_iphone
     get '/', params: {ua: :iphone}
     assert_response :success
     assert_template :"show"
-    assert response.body.include?('show.iphone.html.erb')
+    assert response.body.include?('show.html+iphone.erb')
   end
 
   def test_トップ_android
     get '/', params: {ua: :android}
     assert_response :success
     assert_template :"show"
-    assert response.body.include?('show.sp.html.erb'), 'show.android.html.erb は無いので show.sp.html.erb を表示する'
+    assert response.body.include?('show.html+sp.erb'), 'show.html+android.erb は無いので show.html+sp.erb を表示する'
   end
 end