+++ lib/bizside/user_agent/action_view.rb
@@ -1,15 +1,13 @@
+# TODO: Rails 5 のサポートが終了したらこのファイルは不要になるので削除する
+
 case Rails::VERSION::MAJOR
 when 5
-  if Bizside.config.user_agent.use_variant?
-    load File.expand_path(File.join('action_view', 'use_variant.rb'), __dir__)
-  else
-    load File.expand_path(File.join('action_view', 'action_view_4.rb'), __dir__)
+  unless Bizside.config.user_agent.use_variant?
+    raise 'ERROR: config/bizside.yml で "use_varint: true" を指定してください。デバイスにより View ファイル を切り替えている場合は、ファイル名を Rails 標準の形式に変更してください'
   end
 when 6
-  if Bizside.config.user_agent.use_variant?
-    load File.expand_path(File.join('action_view', 'use_variant.rb'), __dir__)
-  else
-    load File.expand_path(File.join('action_view', 'action_view_6.rb'), __dir__)
+  if Bizside.config.user_agent.to_h.has_key?('use_variant')
+    warn("DEPRECATION WARNING: use_variant is deprecated. Delete it from config/bizside.yml")
   end
 else
   raise "Rails#{Rails::VERSION::MAJOR} はサポートしていません。"